% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_sankey.r
\name{as.stage.series}
\alias{as.stage.series}
\title{Stage series}
\usage{
as.stage.series(x, stages)
}
\arguments{
\item{x}{data.frame; input data}

\item{stages}{a vector of names of columns}
}
\value{
a stage.series object
}
\description{
Stage series
}
