% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_sankey.r
\name{hierarchical.sankey}
\alias{hierarchical.sankey}
\title{Sankey diagram with hierarchy.}
\usage{
hierarchical.sankey(x.ss, bar.width = 10, bar.dist.min = 10,
  xlab = "Stage", ylab = "State", band.col = "green", band.alpha = 0.3)
}
\arguments{
\item{x.ss}{a stage.series object}

\item{bar.width}{width of bars}

\item{bar.dist.min}{the minimal distance between bars in the same stage}

\item{xlab}{label of stages}

\item{ylab}{label of states or levels}

\item{band.col}{colour of flow bands}

\item{band.alpha}{opacity of flow bands}
}
\value{
hierarchical Sankey with ggplot
}
\description{
Sankey diagram with hierarchy.
}
